login method work
this begins the work on the actual user account logon procedure.  three
cases: 1. create unused account when client does not currently own an
account, 2. log in to client's own account, 3. log in to account not
owned by client but linked from another account (whitelisted for log in
by this client) using a password